+++ src/command-selector.ts
@@ -1,6 +1,7 @@
 import minimist from "minimist";
 import { deploy } from './commands/deploy';
 import { proposeUpgrade } from './commands/propose-upgrade';
+import { getApprovalProcess } from "./commands/get-approval-process";
 
 const USAGE = 'Usage: npx @openzeppelin/defender-deploy-client-cli <COMMAND> <OPTIONS>';
 const DETAILS = `
@@ -9,6 +10,8 @@ Performs actions using OpenZeppelin Defender.
 Available commands:
   deploy  Deploys a contract.
   proposeUpgrade  Proposes an upgrade.
+  getDeployApprovalProcess  Gets the deploy approval process configured for a network.
+  getUpgradeApprovalProcess  Gets the upgrade approval process configured for a network.
 
 Run 'npx @openzeppelin/defender-deploy-client-cli <COMMAND> --help' for more information on a command.
 `;
@@ -24,6 +27,8 @@ export async function main(args: string[]): Promise<void> {
       await deploy(args.slice(1));
     } else if (regularArgs[0] === 'proposeUpgrade') {
       await proposeUpgrade(args.slice(1));
+    } else if (regularArgs[0] === 'getDeployApprovalProcess' || regularArgs[0] === 'getUpgradeApprovalProcess') {
+      await getApprovalProcess(regularArgs[0], args.slice(1));
     } else {
       throw new Error(`\
 Unknown command: ${regularArgs[0]}