+++ src/commands/propose-upgrade.ts
@@ -1,10 +1,10 @@
 import minimist from 'minimist';
 import { FunctionArgs, upgradeContract } from '../internal/upgrade-contract';
 import { getDeployClient } from '../internal/client';
-import { getAndValidateString, getNetwork } from '../internal/utils';
+import { USAGE_COMMAND_PREFIX, getAndValidateString, getNetwork } from '../internal/utils';
 import { DeployClient } from '@openzeppelin/defender-sdk-deploy-client';
 
-const USAGE = 'Usage: npx @openzeppelin/defender-deploy-client-cli proposeUpgrade --proxyAddress <PROXY_ADDRESS> --newImplementationAddress <NEW_IMPLEMENTATION_ADDRESS> --chainId <CHAIN_ID> [--proxyAdminAddress <PROXY_ADMIN_ADDRESS>] [--abiFile <CONTRACT_ARTIFACT_FILE_PATH>] [--approvalProcessId <UPGRADE_APPROVAL_PROCESS_ID>]';
+const USAGE = `${USAGE_COMMAND_PREFIX} proposeUpgrade --proxyAddress <PROXY_ADDRESS> --newImplementationAddress <NEW_IMPLEMENTATION_ADDRESS> --chainId <CHAIN_ID> [--proxyAdminAddress <PROXY_ADMIN_ADDRESS>] [--contractArtifactFile <CONTRACT_ARTIFACT_FILE_PATH>] [--approvalProcessId <UPGRADE_APPROVAL_PROCESS_ID>]`;
 const DETAILS = `
 Proposes an upgrade using OpenZeppelin Defender.
 
@@ -15,14 +15,14 @@ Required options:
 
 Additional options:
   --proxyAdminAddress <PROXY_ADMIN_ADDRESS>  Address of the proxy's admin. Required if the proxy is a transparent proxy.
-  --abiFile <CONTRACT_ARTIFACT_FILE_PATH>  Path to a JSON file that contains an "abi" entry, where its value will be used as the new implementation ABI.
+  --contractArtifactFile <CONTRACT_ARTIFACT_FILE_PATH>  Path to a JSON file that contains an "abi" entry, where its value will be used as the new implementation ABI.
   --approvalProcessId <UPGRADE_APPROVAL_PROCESS_ID>  The ID of the upgrade approval process. Defaults to the upgrade approval process configured for your deployment environment on Defender.
 `;
 
 export async function proposeUpgrade(args: string[], deployClient?: DeployClient): Promise<void> {
   const { parsedArgs, extraArgs } = parseArgs(args);
 
-  if (!help(parsedArgs, extraArgs)) {
+  if (!help(parsedArgs)) {
     const functionArgs = getFunctionArgs(parsedArgs, extraArgs);
     const client = deployClient ?? getDeployClient();
     const upgradeResponse = await upgradeContract(functionArgs, client);
@@ -40,20 +40,20 @@ function parseArgs(args: string[]) {
     boolean: [
       'help',
     ],
-    string: ['proxyAddress', 'newImplementationAddress', 'chainId', 'proxyAdminAddress', 'abiFile', 'approvalProcessId'],
+    string: ['proxyAddress', 'newImplementationAddress', 'chainId', 'proxyAdminAddress', 'contractArtifactFile', 'approvalProcessId'],
     alias: { h: 'help' },
   });
   const extraArgs = parsedArgs._;
   return { parsedArgs, extraArgs };
 }
 
-function help(parsedArgs: minimist.ParsedArgs, extraArgs: string[]): boolean {
-  if (parsedArgs['help']) {
+function help(parsedArgs: minimist.ParsedArgs): boolean {
+  if (!parsedArgs['help']) {
+    return false;
+  } else {
     console.log(USAGE);
     console.log(DETAILS);
     return true;
-  } else {
-    return false;
   }
 }
 
@@ -62,7 +62,7 @@ function help(parsedArgs: minimist.ParsedArgs, extraArgs: string[]): boolean {
  * @returns Function arguments
  * @throws Error if any arguments or options are invalid.
  */
-export function getFunctionArgs(parsedArgs: minimist.ParsedArgs, extraArgs: string[]): FunctionArgs {
+function getFunctionArgs(parsedArgs: minimist.ParsedArgs, extraArgs: string[]): FunctionArgs {
   if (extraArgs.length !== 0) {
     throw new Error('The proposeUpgrade command does not take any arguments, only options.');
   } else {
@@ -75,16 +75,14 @@ export function getFunctionArgs(parsedArgs: minimist.ParsedArgs, extraArgs: stri
 
     // Additional options
     const proxyAdminAddress = getAndValidateString(parsedArgs, 'proxyAdminAddress');
-    const abiFile = getAndValidateString(parsedArgs, 'abiFile');
+    const contractArtifactFile = getAndValidateString(parsedArgs, 'contractArtifactFile');
     const approvalProcessId = getAndValidateString(parsedArgs, 'approvalProcessId');
 
     checkInvalidArgs(parsedArgs);
 
-    return { proxyAddress, newImplementationAddress, network, proxyAdminAddress, abiFile, approvalProcessId };
+    return { proxyAddress, newImplementationAddress, network, proxyAdminAddress, contractArtifactFile, approvalProcessId };
   }
 }
-
-
 
 function checkInvalidArgs(parsedArgs: minimist.ParsedArgs) {
   const invalidArgs = Object.keys(parsedArgs).filter(
@@ -97,7 +95,7 @@ function checkInvalidArgs(parsedArgs: minimist.ParsedArgs) {
         'newImplementationAddress',
         'chainId',
         'proxyAdminAddress',
-        'abiFile',
+        'contractArtifactFile',
         'approvalProcessId',
       ].includes(key),
   );