+++ src/internal/upgrade-contract.ts
@@ -8,14 +8,14 @@ export interface FunctionArgs {
   newImplementationAddress: string;
   network: Network;
   proxyAdminAddress?: string;
-  abiFile?: string;
+  contractArtifactFile?: string;
   approvalProcessId?: string;
 }
 
 export async function upgradeContract(args: FunctionArgs, client: DeployClient): Promise<UpgradeContractResponse> {
   let newImplementationABI: string | undefined;
-  if (args.abiFile !== undefined) {
-    const artifactObject = JSON.parse(await fs.readFile(args.abiFile, 'utf8'));
+  if (args.contractArtifactFile !== undefined) {
+    const artifactObject = JSON.parse(await fs.readFile(args.contractArtifactFile, 'utf8'));
     newImplementationABI = JSON.stringify(artifactObject.abi);
   }