+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.0.1-alpha.4 (2024-02-14)
+
+- Add commands to get approval process information. ([#4](https://github.com/OpenZeppelin/defender-deploy-client-cli/pull/4))
+- Renamed options, see breaking changes below.
+
+### Breaking changes:
+- `deploy` command: Renamed the `--artifactFile` option to `--buildInfoFile`
+- `proposeUpgrade` command: Renamed the `--abiFile` option to `--contractArtifactFile`
+
 ## 0.0.1-alpha.3 (2024-02-09)
 
 - Support proposing upgrades. ([#2](https://github.com/OpenZeppelin/defender-deploy-client-cli/pull/2))